refactor: 削除モーダルのコンポーネント分割
